Audio: HAL adds alsa-lib (external/alsa-lib) as part of its library
libaudio is a mapped library, but libasound is not. Non-mapped library
space is limited but the space allocated for libaudio is large enough
to acommodate the HAL and alsa-lib. So, libasound static library will
be added to libaudio.

libaudio now requires of ALSA library from external/alsa-lib, so the
manifest file must include "alsa-lib" project.

Signed-off-by: Sean McNeil <sean.mcneil@windriver.com>
Signed-off-by: Misael Lopez Cruz <x0052729@ti.com>