Fixes for several issues
Issue: OMAPS00204838: When a video is opened in the gallery app, it displays an error message.
Resolution: The issue was only wrt h263. It hung in stream off after resizing the input. Implemented a workaround (set controlReady to
false after resizing) in overlay module.

Issue: A bug was introduced when we rewrote writeAsync which caused one of the buffers to be lost. Only 3 out of the 4 buffers were being
used.
Resolution: This issue was fixed. We were returning the wrong cmdid.

Issue: Camcorder - recorded video was jerky as if the frames were repeated.
Resolution: The issue was in CameraHal buffer handling mechanism. Fixed.

Wrote an application to test uncovered SRs. Tested all SRs currently supported by DSS. 6 out of the 9 tests moved from U to P. The
remaining 3 tests (global alpha, ARGB8888, ARGB4444) are yet to be supported by the DSS.

Updated the copyright headers in liboverlay/*.

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>