Fix zoom issue and add new configuration for IPP
The zoom was not been updated from the camera application, with these changes the zoom works ok but when it is set for 6x or 7x there are artifacts with some resolutions. However accoarding to the SRs we should provide support for 1x,2x,3x and 4x.

The new IPP configuration follows the next sequence:
YUV422I->YUV42O conversion
EENF or CS algorithm
YUV42OP->YUV422I conversion

To enable this configuration next flags should be as follows:
	#define IPP_YUV422P 0
	#define IPP_YUV420P_OUTPUT_YUV422I 1