Fix for multiple nal units per frame
Dynamic framerate exposed a bug in OMX as the algorithm started producing
frames with multiple NAL units. OMX was tagging correct time stamp only
to last NAL unit in each frame. This has been fixed now to tag the same
timestamp to all NAL units belonging to each frame. Looks like PV muxer
and demuxer are compatible and apparently this is not an issue for them.
But quicktime and VLC don't like NAL units with 0 timestamp in the
middle of a bitstream.

Originally from: https://partner.source.android.com/g/#change,1402