Remove (1000 ms-long) unnecessary waiting time
The change in the MessagingThread.
An event is used to indicate the readiness of (at least one) message(s) in a message queue
Only after an event occurs, DSPNode_GetMessage is called to retrieve a message from the queue.
When retrieving the messages, if the queue happens to contain no messages at all, the call (DSPNode_GetMessage) will be blocked
for the given amount of timeout period.
The current logic seems to always block for the given amount of timeout period, and use that to indicate there is no message
in the queue.

This is unnecessary, since when there is no message, DSPNode_GetMessage returns an error anyway; and when there is some message,
the call is not blocking.

Thus, we should remove the wasteful waiting time.

bug - 2414536