Camcorder: Fix preview freezing after multiple captures
Sometimes, when recording is stopped, opencore does not
return all the buffers. When we do multiple recordings, the
buffers get lost gradually and causes the camera to starve
- which results in preview freezing. Logic has been added
to reclaim all the buffers when stopRecording is called.

For higher bitrates, the VE takes longer to encode.
Therefore all the buffers are with the VE and the
camera begins to starve. The preview looks very jerky
under those conditions. To alleviate the situation,
if the camera is starving, then dont send the buffers
to VE and instead queue it back to the camera.

Change-Id: I97f9da5d3863e0ba415ca3e0f54bd3052a7917db
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>