Camcorder: Use V4L2 buffer timestamps
Until now, the video timestamps were nothing but the system time when the record callback was called.
It is more accurate to use the timestamp populated by the Camera driver in the v4l2_buffer structure.
This will also eliminate the variations seen in the timestamps This should also provide better AV Sync.

Change-Id: I19131e2e7613e0501984dd92b571177a0c186eb2
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>