OMAP4: opencorehwlib surfaceoutput chnages for co-existance of ducati+arm codecs
these are the chnages for separate shared library for pvcore surface output.
platform specific surface output is modified to add checks for video format.
the check is based on the mime string passed from pv core.
this is required to decide whether color conversion is required or not.
so that overlay can be created as per the format given by ducati codec or arm codec.

Change-Id: Idd1753e2dcff28e66c5c4a242830117214a2f1ce
Signed-off-by: sunita <sunitan@ti.com>