IMAGE: OMAP4: Libskiahwdec implementation for JPEG Decoder.
This code is Initializing, configuring the OMX functions required by
our hw, taking care of OMX state change, allocation and freeing memory.

At the end this code should run with the execution of Gallery and "SkLibTiJpeg_Test”
Test, but for being able to run, it requires to be enable enable with:
"setprop jpeg.libskiahw.decoder.enable 1" command.

This patch is dependent of: Ic820ea669ea480e5a2319426e3fbdfd0ebdc2066, before
submitting it, this patch should be merge.
Signed-off-by: Octaviano Flores <x0083028@ti.com>

Change-Id: If88583f73d36ff78dd4f07276eb254743479fade