CameraHal: KW defect 1929
Buffer overflow, array index of 'tmpBuffer' may be out of bounds.
Array 'tmpBuffer' of size 513 may use index value(s) 512..513

Buffer overflow could be possible due to strncat using buffersize
as length and not bounding it dynamically. The fix is to replace
strncat with strlcat, which is buffer overflow safe. Added additional
error checking for truncation. Right now we will just exit in case
of truncation. If the truncation condition is ever met, the exit
might cause an issue at runtime.

Change-Id: I0c7693f581d3eac52730717d3611eb23e1f602e6