OMAP4: Fix for starting preview conditionally inside startPreview() and other misc fixes
This fix makes sure that we start the preview inside startPreview() only if either the application
has registered a valid preview callback or a valid overlay has been set. If none of these conditions
are met, then preview is not started. It will be started later when either of these conditions become
true.

Note however that the condition to proceed with starting the preview with app callback alone has
been disabled and preview will start only after a valid overlay is provided.

This is because of a OMX_FlushBuffer bug in the ducati implementation of OMX_Camera which prevents
us from switching to preview buffers once the overlay becomes available. Once this is fixed in
ducati release L24.6, we need to re-enable this check.

Change-Id: I47aee71d3bb037e8d69142dd7bebeada0a9ab4ac
Signed-off-by: Sundar Raman <sunds@ti.com>