Overlay and CameraHal: Handling multiple stream off/on
DR: OMAPS00217786

Issue: The camcorder preview becomes black when this sequence is executed:

Open Camera App
Switch to Camcorder mode
Press the HOME key and exit app
Open another app like web browser, BT file transfer, Gallery or Alarms
Come back to Camcorder app
Now, there is no preview. Its black.

Root cause:

When the above sequence is executed, for some reason, the surface flinger
changes the window position multiple times. In overlay, when the window
parameters are applied, the stream is disabled. When the stream is disabled,
overlay discards all the buffers queued to it. So, cameraHal cannot dequeue
those buffers. This is not communicated to CameraHal. So, CameraHal cannot
queue the buffers back to Overlay. The stream will be enabled only after at
least 2 buffers have been queued to overlay. So, we are in a deadlock. Since
stream is disabled, the preview is black.

Changes have been made to Overlay and CameraHal so that a streamOFF may
be inferred by the cameraHal and it can take appropriate action.

Overlay:

1. The overlay_queueBuffer function has been modified to return the number
of buffers queued to the DSS at that point. If there is an error, it returns a
negative value. But if there is no error, then it returns the number of buffers
queued to the DSS. Locks are used to avoid race conditions.

2. The method overlay_dequeue has been modified to return the error code -2 if stream is off.

CameraHal:

The CameraHal keeps track of how many buffers were queued to overlay.
If this does not match the count returned by overlay_queueBuffer, then it
implies that a 'stream off' happened in overlay and so the buffers discarded
by overlay are queued back to camera.

libopencorehw:

Adapting to changes in made in Overlay( return values of queueBuffer and dequeueBuffer).

Change-Id: I8db977d290ac0a1aeb6205f0a25e41f73c52ca68
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>