MAP4:Overlay max size limit for omap4.
Desc: In order to avoid the error scenarios where arbitrary overlay sizes
are requested, the max limits are checked for the overlay creation. the constants
are defined for omap4 as 1080p with some buffering to take care of padded buffers.

validation:try to create a overlay with the size beyond 1080p from overlay_test.
the overlay creation should return error code, without faulting the system.

Change-Id: Ia7372eb1d5915884cc50286a3c168f371df877a1
Signed-off-by: sunita <sunitan@ti.com>