OMAP4: Overlay class based re-arch for single overlay
Desc: In order to support the advanced features of Overlay and the multiple
overlay objects in OMAP4, the existing OMAP3 overlay data structures are
re arranged into different objects so that the number of data structures
can be reduced and some of the data copies can be avoided.
This patch is to support the single overlay with the modified architecture.
the interface to the Overlay still remains the same as in OMAP3, only the
internal architecture has been modified. Hence there is no impact to the
Overlay clients like Video MIO and CameraHAL. This submission is the basis
for the subsequent patches for overlay advanced features and multiple overlay
objects.

Validation: Validated overlay component test app and video playback from
Android media gallery.

Change-Id: Id3d9caece220ce968d6bfae8b8a4db844964756c
Signed-off-by: sunita <sunitan@ti.com>