OMAP4:Multiple Overlay Objects support
Description: This submission is to support the multiple instances of the overlay
objects simultaneously. The maximum possible overlay instances are defined
as 3 for OMAP4 and as 1 for OMAP3. The actual number of available overlay
instances depends on the number of framebuffer devices configured in the
system.

Validation: Validated two overlay paths simultaneously on the primary LCD.
one overlay is being used by the media player and the second one is used
by the overlay component test app to render the raw video frames.

Change-Id: If5fa8850bfa005cc9cf39483334b67f89c00ef59
Signed-off-by: sunita <sunitan@ti.com>