MAP4: Overlay changes to support dynamic assignment of video pipeline to the overlay object
Description: The number of video pipelines available for the overlay object
varies based on the number of framebuffer devices reserved in the system. Hence
the video pipe line should be queried run time and assign to the overlay object.
so that based on the FB devices in the system, the remaining video devices would be assigned to the overlay devices at run time.

Validation: run overlay_test app to create the multiple overlay instances.
enable the traces in Overlay and observe that different overlays are assigned
with 1 fb device and with 2 fb devices in the system.

Change-Id: If512b991c444e6ab1addd4e5fd779faefd810ffd
Signed-off-by: sunita <sunitan@ti.com>