OMAP4: Overlay chnages to support alpha, zorder and color key settings
Description: Overlay library setParameter interface is extended to support
the dynamic configuration of global alpha, zorder and color key settings
for the current overlay and manager. And the video transparency feature is
enabled with the default zorder and the transparency key values.the alpha
and Zorder settings are kept under OMAP4 macro as these are specific
to OMAP4.

Validation: run overlay_test app from command line, select options for alpha
testing and zorder testing. Enter the values for alpha and zorder from the
terminal. observe the chnages in the video frame alpha and zorder properties.

Change-Id: I766a27dea68d3a907538e8c1729c691f0647e1cf
Signed-off-by: sunita <sunitan@ti.com>