OMAP4:Overlay fix to update the new width and height in resizeInput
Description:The current overlay resizeInput function doesn't cache the new width
and height of overlay. This causes the next resizeInput functionality to
fail if it passes the width and height same as the initial w &h. this is
fixed by updating the width & height of overlay object in resizeInput method.
so that any number of resizeInput will work from any valid size to any other
valid size.

Validation: executed overlay_test app to do resize multiple times for different
sizes.

Change-Id: I4ee49865893e7a4bc498a2371806df997b524890
Signed-off-by: sunita <sunitan@ti.com>