OMAP4: Overlay support for dynamic configuration for number of overlay buffers
Description:Current overlay library defines a constant for the number of overlay
buffers to be used. This constant may not be optimal for all the clients like
Video MIO and Camera HAL might have different requirements on the number of
buffers they have to use. And creating the highest number of buffers for all
clients is not acceptable from memory consumption point of view. Hence, overlay
setParameter interface is extended to support the client to configure the
number of overlay buffers required. and the existing constant would be used
of the client doesnt configure the number of buffers explicitly, and is also
used as the max limit for any client.

Validation: Executed overlay_test component test app to request for
less number of numbers than the constant defined in the overlay library and
observe from the trace log that the requested number of v4l2 buffers are created.
And request for higer number of buffers than the overlay constant, and observe
that, in this scenario client request is ignored and overlay default number
of buffers are created

Change-Id: Iadf36330c497b32d981e17146d729f7c08875baf
Signed-off-by: sunita <sunitan@ti.com>