OMAP4: Overlay changes to support HDTV selection from command line
Description: The current overlay design supports the panel configuration based
on the "screenId" requested from the surface flinger. In the complete solution,
each applicatio has to set the sceeen ID while requesting the window manager
for the surface creation. Alternatively we can configure the HDTV and set
the overlay manager to "tv" from the command line via sysFs interface.
This patch is the support in overlay to read the overlay mananger via sysFS,
set from the command line to the driver, and adjust the window size for HDTV.

Validation: Tested 1080p and 720p Videoplayback on HDTV, with the HDTV enabling
and overlay manager setting from command line sysFS interface.

Change-Id: I290ee721f9b36ffc14e79859f61a855b988f2568
Signed-off-by: sunita <sunitan@ti.com>