OMAP4: Overlay window calculation for HDMI and LCD
Description: The window size for HDTV is calculated based on the Video frame
width and height instead of hardcoding it to 1080x1920. this is required
(a) to maintain the aspect ratio of the video
(b) to avoid odd resoultion clips getting unnecessarily scaledup

Validation: Tested 30 AV clips for playback on HDTV and LCD displays. this
covers different resolutions and lanscape/portrait clips. All clips are
playing fine on both LCD and HDTV.

Change-Id: Ida4a195409fa908df628af18bb876238c0a0cc73
Signed-off-by: sunita <sunitan@ti.com>