OMAP4 Overlay changes to initilize Crop & Window parameters to non Zero values
Description: In order to maintain the aspect ratio of the video, window size
is calculated based on the cropped video size. Crop values are set from the
overlay client process which is video MIO or the Camera service, where as
the window calculation is triggered from the surface flinger process.
Hence there is a race condition in using the crop values even before they
are set from the Video mio / Camera service. If the crop values are zero,
the window size would become zero, from the math, and hence faile in v4l2.
In order to fix this, the crop width and height are initialized to the overlay
width and height by defualt and would be chnaged by the Video MIO / camera
service later.

Validation: Validated AV playback on LCD and HDTV. All resolutions are working
fine.

Change-Id: Ibba29e9d2413aaa1b9c88a40dc665fc60ad7c864
Signed-off-by: sunita <sunitan@ti.com>