OMAP4 CameraHAL: Remove polling on V4L2 fd and use optimum number of buffers with display to reduce ARM load
This patch removes the polling of the V4L2 fd for getting to know when to DQ a buffer. Currently, we poll the fd and if there is buffer to DQ,
we then check if we have reached the optimum buffer count. This results in continuous loop executing and checking for this condition.
Now, we are checking if the number of buffers with display has reached the optimum number of buffers and only then post a message to the
display thread to attempt a DQ. This way the continuous loop is eliminated and the load on ARM is reduced.

Change-Id: I53697f90dca9b9a4680866bd8a7452d89be150c0
Signed-off-by: Sundar Raman <sunds@ti.com>