OMAP4 CameraHAL: Support for displaying padded buffers when any of LDC-NSF,VNF or VSTAB is enabled
This patch provides support for displaying the valid frame data existing within a larger buffer
as output from the LDC-NSF, VNF or VSTAB algorithms. The nOffset field part of the OMX header
is utilized to calculate the x and y offsets for the start of the valid frame, and setCrop API is
used to convey these offsets to the Overlay.

Validated YUV422I and NV12 HQ preview with this patch

This patch has dependencies on 3837 & 3834 Overlay patches

Change-Id: I028e01fc37b64c6b8e17d4ba59941ee0821a268a
Signed-off-by: Sundar Raman <sunds@ti.com>

Eclair port.