OMAP4:Overlay on-the-fly Crop support for VSTAB usecase
Description:In case of Video recording with VSTAB enabled, the video buffers
will have different offsets for the stable frame with-in the buffer. And
this offset changes from buffer to buffer. Inorder to achieve the proper
stable video preview, we have to set the buffer crop co-ordinates on-the-fly
without disabling the streams. Overlay setCrop() API is modified to check
if the setCrop is only changing the crop co-ordinates, keeping the crop
window same. In this scenario, v4l2 setcrop is called without first disabling
the stream.
Modified the overlay_test app to test this feature.

Validation: Tested overlay_test component test app for continuous crop.

Change-Id: I5db39fc39a9c3a064163e0bf22b4b43741b050fc
Signed-off-by: sunita <sunitan@ti.com>