IMAGE:LIBSKIAHW-OMAP3: Fix for a minor DR-OMAPS00216669.
[Comment]
In omap3, EmptyBufferDone callback is always first. Suppose if
FillBufferDone callback comes before EmtyBufferDone callback there
will be a hang issue in the jpeg decoder in libskiahw.
This scenario faced in OMAP4 testcases with the ducati OMX
implementation.

Eventhoguh there is no issues in omap3 till now, it is better to
check for both callback state when move to IDLE from EXECUTING
state in libskiahw-omap3/SkImageDecoder_libtijpeg.cpp.

This patch will do that check.

[DR]
OMAPS00216669

FROYO VERSION OF 3204

Change-Id: I165a1ee8e1356291bb4e426bcdf768dbba4c7198