IMAGE:LIBSKIAHW: Fix for DR OMAPS00218395.
[Comment]
Issue:
Sometimes, the thumbnail is not displayed, and [ ! ] is shown on
the gallery application for the images bigger than 12MP even if
it has the embedded EXIF thumbnail.

Root Cause:
This is due to the timing issue as the libskiahw took more time
to exit from onDecode(). Since it loads the OMX and DSP codecs
and then it returns due to >12MP from OMX setParameter.

Solution:
Check for the image resolution in onDecode() function and return
immediately without loading the OMX and DSP codec.

This patch will fix this issue.

[DR]
OMAPS00218395

FROYO VERSION OF 3378

Change-Id: I8c77ad227989c73e99bd7e07d2b34f0e251100fb