OMAP4: Audio: avoid to set up an audio route when no device is given
The method s_route can be called with no devices specified by the audio policy manager.
The method is not supposed to handle this kind of audio routing.
To avoid this problem audio routing will not be changed if no device is given.
Crash was observed when starting new voice call.

Change-Id: I694de4ed3befa612e3d392c2c9b855d4daddbbd4