LIBSKIAHW OMAP3: JPEGs with large header not decoding properly
Root Cause:
Buffers passed to libskiahw via BitmapFactory use BufferedInputStream
to pass the JPEG stream. BitmapFactory has set the readlimit to 1024
bytes. After reading 1024 bytes, BufferedInputStream is invalidated
and user cannot rewind back to the beginning of the stream. Original
implementation of libskiahw requires reading from the stream twice,
once to parse the JPEG header and another to read entire JPEG stream.
If we read passed 1024 bytes for the JPEG header, we cannot rewind to
the beginning to read entire JPEG stream.

Solution:
This patch allows libskiahw to only have to read from the stream once.

Change-Id: I437efa54d4ba4d4ea80e55488f13d8a0bf1b2878