OMAP4: TIHardware Renderer changes to support padded&offsetd video buffers
Description: TI Hardware Renderer is enhanced to handle the padded buffers,
and the buffers with offset from base address. Based on the offsetd address
received from AwesomePlayer, the buffer index is identified and the cropX
and cropY values are calculated based on the offset.
And the overlay is configuired with the calculated cropX, cropY and crop Window
dimensions.

Validation: Tested Video playback from Gallery.

Change-Id: If63b71960a94a317713eab17809c30a99c28d9e8
Signed-off-by: sunita <sunitan@ti.com>