CameraHal OMAP3: Set length of overlay buffer when queuing to camera
CameraHal was ignoring the length of the overlay buffer when queuing
it to the camera driver. Camera driver could go out of bounds when
writing in the overlay buffer if camera buffer is larger than
overlaybuffer. Other potential issues could occur if the buffer
sizes are not the same. Now we set the length of overlay buffer
when queueing to camera driver, and camera driver can determine
how to handle this case.

Change-Id: I0c6c9ea826e86b34f5d792e1356571416931b84a