OMAP4 Camera HAL: Fix for AF timeout check
In the checkFocus method, we are currently not checking for the focus errors passed by Ducati and instead we simply timeout after 5 seconds.

This patch listens for the following errors - OMX_FocusStatusOff, OMX_FocusStatusUnableToReach and passes them to the application.

With this patch, we are able to proceed with image capture within 200-300 mS rather than the 5 second timeout earlier.

This patch is still required even after AF Callback is implemented.

Change-Id: I825eb0267a90183bd2255e42f9500b241563a98a
Signed-off-by: Sundar <sunds@ti.com>