OMAP4 Camera HAL: Support for sensor selection inside CameraAdapter
This patch adds a new property to TICameraProperties.xml called sensor-index, which will
allow the CameraAdapter to choose a specific sensor after instantiation. In other words, it
defines the mapping between logical camera index to physical camera index, specific to
a camera adapter.

This can be used to select the secondary sensor index value on Ducati side for secondary
camera logical index.

Change-Id: I715ba83927f106813a4a180f87915c03f55d9271
Signed-off-by: Sundar <sunds@ti.com>