CameraHal OMAP3: KW 11093
Double freeing of freed memory may be in class 'CameraHal'.
Copy constructor is not defined.

ICaptureDestroy could be called on a "copy" (shallow) of
CameraHal class. In this case, ICapture structures could be
double freed. Adding if statements around memory being
freed to guarantee that the memory will only be freed once.

Change-Id: I31df840b9cd8666bb5fc2095fef3e1456142ebe1