Fps drop seen when VSTAB / VNF is turned on the fly
There was fps drop seen in turning on the VSTAB / VNF on the fly since
we were closing the camera and then turning on the preview on. This is
now properly done in application by restarting the preview. Side effect was
seen in VSTAB turning on since the enabledisplay function use to have mWidth
and mHeight as zero. A safe gaurd read is added to avoid reading zero.

Change-Id: I359b9fbf78c9488629648b7b43b230d13c05ccbe