IMAGE:LIBSKIAHW-OMAP3: Exit jpeg handle gracefully during timeout.
[Comment]
This patch will take care of pushing the omx-jpeg in to invalid
state before the libskiahw starts freeing the jpeg handle and
freeing the memory.
This was not taken care in the situation when the watchdog timer
expires and the watchdog callback initiates the de-initialization.
This patch is created while analysing the DR OMAPS00222328.

[DR]
-

Change-Id: I1de132e8b81aac912d5a513b11b06522c21fe105
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>