DSPBRIDGE: use and propagate the correct errcode to the user
Modify the bridge's interface and APIs to propagate the errno
to the user. Old error codes included in errbase.h are deprecated
and the file was removed; Macros related to the propagation of the
error codes were also removed. Please note, DSP_STATUS was removed
and now Bridge's APIs return int data types.
MM must take care of the below changes:

- DSP_SOK is deprecated, 0 must be used.
- errbase.h was removed.
- DSP_STATUS is not longer available. int must be used.
- kernel error codes will be propagated to the user.

Change-Id: I952f236bcfcf7e3060c9cb00e7d1fbb07f234f8e
Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>
Signed-off-by: Rene Sapiens <rene.sapiens@ti.com>