IMAGE:LIBSKIAHW TEST APP: JPGD type(SW/DSP) check updated.
[Comment]
This patch updates the logic of choosing the JPGD type as
implemented in DFactory() in skia layer. This is to avoid
setting the SubRegion parameters if the JPGD is SW decoder.
This patch also corrects the FAIL test case counts.
This patch also updates the md5sum value for one test case.

[DR]
-

Change-Id: I1ad84ea162f35340b13143adccc9af8bcbd7d616
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>