hardware/ti/omap3: introduce new uim-rfkill
On 2.6.35 the TI shared transport driver moved to the rfkill way
of exchanging notifications, unlike the old sysfs-entry/PID was on x.32/x.34.
This daemon opens/polls and reads on the /dev/rfkill, waiting for
notification from the shared transport driver to either install/uninstall
the line discipline on the UART provided in the command line.

It also does other smaller things like changing the permissions of the rfkill
entry to allow BT to be turned on by the application with UID bluetooth.

Change-Id: I77dcb7c8cbaaf6565b896f9aa02a30ecac30217a
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>