OMAP4: Overlay workaround for .35 kernel
Description: Since some of the custom V4l2 methods are missing in .35 kernel,
overlay library is updated to not use those apis and the references to the
picoDLP are removed till it is supported in kernel.

All the changes are kept under KERNEL_35_WA macro, defined only for OMAP4.

Change-Id: Icc0ffeaa9d0c1a97be7d8124b4a6871e8298816c
validation: Tested overlay_test component test app on latest FS + .35 kernel