Removing potentially misleading print messages.
Changing LOG level of some messages to "LOGV" so they are not printed by default.
These messages are causing some confusion in the daily builds because they sound
like an error occurred even though the appearance of these log messages is not
necessarily bad.  If these messages are still desired, they can be enabled by
logging in verbose mode.

Change-Id: I5e1e3f278513bd29f3cdc9c5ef7da6a5593a26d8