OMAP4: Added FM Rx mixer support via ABE for OMAP4
FM Rx on OMAP4 is via the ABE loopback path.
The capture is done via the PDM-UL with source being Aux-FM,
via the AMIC-UL, down to DL1's SDT mixer to be played back
on the DL1 PDM via the HS/HF.

what this patch does:-
When the device/route is set to be FM via the FM application,
sets the relevant mixer settings.
It also opens/configures the playback and capture devices to
start the pcm.

This depends on the host-less transfer implementation to avoid
application doing a specific arecord or aplay.

Author:    shitiz <shitiz_kumar@ti.com>
Change-Id: If83e1ac7c054640154bc4eaf78175fcf9ed8f4c3
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>