OMAP4: Primary Camera crashes in Video Mode preview
Description: Up on switch from Camera mode to Video mode, Camera preview
crashes with segmentation fault and leaves black screen.
The issue is due to a check for 3D stereoscopic preview in startPreview.
The parameter is valid only in case of seconday/stereo camera's, and is
not populated in the Camera properties xml file for the primary camera.
Since the camera properties file returns NULL for this paramater, the string compare
is causing segmentation fault while trying to access this.

It is fixed by checking for non-NULL before accessing there s3d parameters.

Validation: Tesetd Camera preview in both Camera mode and Video Mode.

Change-Id: Icb79e8c697f2cadecc94993f80a3d594d452f164
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>