OMAP4 libOverlay: Add timeout to DQBUFs
Now dqbufs time out within 66ms to avoid hanging the system.

Note that later this will be reduced to 0ms in most circumstances.

Signed-off-by: Lajos Molnar <molnar@ti.com>

Created a new patch from http://review.omapzoom.org/#change,publish,8033,2
as that patch was corrupted in the server.  This patch is unchanged. I am
taking verifications and approvals from 8033/2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Change-Id: I9d2140418e1243ce156d340f3b6044989bb3b577