hardware/ti/omap3: uim: build for release
uim build Android makefile had 2 flags LOCAL_MODULE_PATH and
LOCAL_MODULE_TAGS which placed the uim executable in
/system/xbin/
and built the uim with engineering module tag.

These are now removed based on a customer enhancement
request.
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>

This was previously patch 8157/1 but that was corrupted in
the Gerrit database.  I created a new patch to enable
merging.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Change-Id: I61e4f82c494511682f8828d1356c2343badc3518