[DRT fix] Cropping needed when rotation is performed
OMAPS00228734 -- Rotate AV Playback shows screen misalignment
                 on first play of clip.

Introduction of patch "http://review.omapzoom.org/#change,7682"
lacked of logic when rotation is performed beforehand.

According to driver, since rotation is to be changed, it is
needed to do a S_CROP in order to adjust panel dimensions
according to the rotation settings, if this is not done,
when change is performed to 90/270 degrees rotation, the video
window dimensions will be limited to 800x480 instead of a
480x800 which causes to get 480x480 instead of a 480x800
window.

Change-Id: I10af5f0d18d7f27adfa3f9f91c95a405ba580afb
Signed-off-by: Omar Torres <otmorales@ti.com>