TIHardwareRenderer changes to support dynamic Resize
Description: For Overlay based Hardware Renderer, media server can't destroy
and create the Renderer locally without destroying the associated Surface.
This is because the overlay object is shared between Media server process
and the SurfaceFlinger process. Hence, if the Media server destroys renderer
and re-create independent of Surface in surface flinger, the overlay re-creation
fails. This kind of dynamic destruct & reconstruct is required to handle
port reconfiguration scenarios.
In order to support this usecase, the TIHardWareRenderer is extended to support
Dynamic resize without actually destroying the overlay object.
This patch provides the support, but the actual logic to call dynamic resize
is in the subsequent patches.

Associated DRT #
OMAPS00227653: Clip H263_500_AMRNB_12.3gp does not display correctly
on the screen

Validation: tested AV playback from Android Gallery

Change-Id: I1a5af751bd3528e7c07d1acab2ba0623d639c222
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>