fix for random AV playback segfault  during system suspend
Description: In the scenario when DQ fails, the debuffer call returns
without populating the buffer index in the caller.
Since in the TI HardwareRenderer, the buffer index is an uninitialized
local variable, sometimes the index is inited to negative or out of bounds,
which is purely run time dependent. In such scenarios, there is a segmentation
fault from hardware Renderer, since it tries to access data structure with the
un-initialized index.
This is fixed by returning an error from lib overlay in DQ failure, hence
Hardware renderer doesnt process any of the local data structures.

Validation: tested AV playback with suspend/resume and Camera preview with
suspend/resume.

Change-Id: I0a6b23fea7b35c0f614ab9129f18164e22b34f80
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>