OMAP4: ALSA: skip the unnecessary function pointer
Since we do not require a specific function for each handle
we can remove the function pointer and call setAlsaControls
directly as needed. This provides the added benefit of getting
easy access to the handle while we are determinig which controls
need to be set.

Note that handle->modPrivate is no longer required, but it is set to
the address of the setAlsaControls function as a safety incase it is
used in future by an external component.

Change-Id: I7dc7c8194066a138243c477ac3913628fb2c3845
Signed-off-by: Chris Kelly <c-kelly@ti.com>