OMAP4 CameraHal: OMXCameraAdapter blocked after canceling image capture
If we cancel in the middle of a HQ image capture, OMXCameraAdapter
is blocked waiting for a ShutterCallback event from Ducati. (See
the "eventSem.Wait()" statement  in the startImageCapture() function.)

This patch unregisters the shutter callback and sends a dummy signal
to the waiting semaphore in startImageCapture() before canceling
the capture.

Change-Id: If080f02526afc4eb2eae5de13f2c68d444500373