libstagefrighthw: Remove sleep of 1s to improve gallery transition time
Currently when a video is played through gallery and it is finished,
gallery is taking long time (~2s) showing blank screen during the
transition. This is mainly due to the 1s sleep in destructor of TIHardwareRenderer.
Even if we remove the sleep there is no issue with overlay destruction.
Hence it is removed to have a smoother transition in the gallery application.

Validated AV playback through Gallery application.

Fixes DR: OMAPS00232501

Change-Id: I61baffec6ee4b3c16c927b7b24b24cd364a432a0
Signed-off-by: dnkmohan<dnk.mohan@ti.com>