OMAP4 CameraHal: Video format parameter is sometimes corrupt
Strings coming from CameraParameters are derived from
the String8 class which uses a shared buffer. After performing
another operation with CameraParameters, the state of shared
buffer is unknown. Therefore, we have to get the value of
preview format again here because the memory pointed to by valstr
is now in an unknown state since we did a set operation which
will allocate from String8 class and could potential invalidate
memory pointed to by valstr.

Change-Id: Ie64b41db28dba85972724a51f19f92aa9b79cc2f