OMAP4 CameraHal: Cancel autofocus before stopping preview
If we exit application in the middle of focusing, OMXCameraAdapter
will eventually timeout waiting for focus event from Ducati. When
the timeout occurs, OMXCameraAdapter will send a dummy callback
to the application. This will sometimes cause a segfault if the
application is already freed. There is also another situation where
we will get an erroneous callback if we start the application
in the 5 second window before OMXCameraAdapter is destroyed.

Change-Id: I40499104a203e8546c5c4e7d242740c7d7c4c7f0