OMAP4:TIHardwareRenderer changes to use codec recommended buffer count for overlay
Description: TIHardware Renderer is modified to receive the number of buffers
to be created by the overlay. The constructor is overloaded to receive
the extra parameter from the awesome player. This patch also involves
the removal of lower limit for the number of buffers. Hence, the number
of overlay buffers requested is purely dictated by the codec.
These is also option to Use number of output buffers as recommended by user/codec.

This patch has to be merged along with
http://review.omapzoom.org/#change,11316

Validation: tested AV playback from Android Gallery app for files of
different resolutions.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch set 2: Cherry-pick on current tree.  No changes made.
Taking verifications, approvals and build test from patch set
1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Change-Id: I86c6fe63c72fb57fe678f77ffcaa02a4bf93d969