Created a new class TIOMXCodec.
It is derived from OMXCodec. All interface changes must be made in this class instead of OMXCodec.

AwesomePlayer must be modified to create an instance of this class instead of OMXCodec.

This is necessary to maintain flash compatibility.

Note# this patch has been ported from Froyo
http://review.omapzoom.org/#change,8893

Patch Set 4: Resolved build issues.

Change-Id: Id6ddb6675e69288035fd878c2f417ba7f2383b25
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>