libOverlay changes to fix Video rotation issue on HDMI
Description# This patch addresses two issues
(1) Since rotation is not applicable for HDMI, the rotation is ignored
for the video device that is handling HDMI panel

(2) In case on Video Cloning usecase we need to have video rotated on LCD
but not HDMI, this is fixed by correcting window settings for cloned device
for LCD.

This patch is tested without UI Cloning

Validation# Tested AV playback cloning on LCD and HDMI with rotation.

patchset2# enabled 180deg Video rotation for HDMI to support video mirroring.

Change-Id: Ie86afbd74f8a7952674f17d9d91284e55025803b
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>