OMAP3:LIBSKIAHW: Cleanup of open source code copy.
[Comment]
This patch removes the files which contains copy of open source
lines.
The required changes have been done to keep this code as TI
proprietary code.
This patch fixes the BlackDuck tool findings and fixes the below
DR in libskiahw-omap3 folder.

This is porting of change #9631 from froyo to gingerbread.

[DR]
-

Change-Id: Ib584950a8e0bc3bd0af3ae9f7d75538db5ca5271
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>