LibOverlay: Redundant logic for window calculation hes been removed
Description# With the addition of new features like rotation,
the logic is made redundant in window calculation for primary
panel and the Cloned panel. This patch removes such redundant logic
and unified all the window calcualtion logic into single method.

Validation# tested AV playback on LCD and HDMI in normal and
rotated mode.

Change-Id: If5c915203ddd755d914ade98a18a38bbb8272ea4
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>