CameraHal OMAP4: AppCallbackNotifier causing deadlock in CameraService
When AppCallbackNotifier is stopped, there is still a chance that
a frame provider will still fill mFrameQ before the frame provider
is stopped. In this case, go ahead and pop the frame off the queue
but just ignore it and return. If we don't do this, there is a chance
that mFrameQ is not empty when AppCallbackNotifier is restarted. In
this case, notifyFrame calls the preview callback prematurely and
causes CameraService to deadlock since CameraService is holding
a lock in startPreview and now waiting on a lock in dataCB.

This patch fixes android.hardware.cts.CameraTest#testPreviewCallbackWithBuffer
in latest CTS (2.3 R2)

Patch Set 2:  Simple cherry-pick, no changes made. Taking verifications,
approvals and build test from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ie6eb89788bf85fd0a016c3b3574475000d438073