TIHardwareRenderer: Fixed to make rotated overlay when it needed.
Added argument rotationDegrees in TIHardwareRenderer::TIHardwareRenderer()
for receive rotation degrees of video clip. and, unified -may havily be
oveloaded- fucntions by gave default value to args.

In this case, the rotated video not means images encoded in potrait size
but, images be encoded as same as taken by camera sensor and have rotations
info. in video track header.

TIHardwareRenderer pass the degrees to liboverlay and the final image of
video be rotated by power of overlay engine.

Change-Id: Iba26fa61e6a8d2541b999d43318f61c0bb4c3e9b
Signed-off-by: Homin Lee <homin.lee@ti.com>